FileDetailsクラスを作成しました
